+++ lib/generator/index.ts
@@ -10,6 +10,7 @@ import Project, {
 	TypeParameterDeclarationStructure
 } from 'ts-simple-ast';
 import { lcFirst } from '../../utils/utils';
+import { assertDefinedValue } from '../../utils/assert-defined';
 
 interface Temp {
 	className: string;
@@ -57,7 +58,7 @@ export default async function generate(source: string, options?: any): Promise<v
 
 		if (generateMethods.length > 0) {
 			generateClasses.push({
-				className: classDeclaration.getName(),
+				className: assertDefinedValue(classDeclaration.getName()),
 				methods: generateMethods.map((generateMethod: MethodDeclaration): MethodDeclarationStructure => {
 					const method: MethodDeclarationStructure = {
 						name: lcFirst(generateMethod.getName()),
@@ -88,12 +89,17 @@ export default async function generate(source: string, options?: any): Promise<v
 						});
 					const preparedParameters: ParameterDeclarationStructure[] = params.slice(1)
 						.map((param: ParameterDeclaration): ParameterDeclarationStructure => {
+							const type = param.getType();
 							const p: ParameterDeclarationStructure = {
-								name: param.getName(),
-								type: param.getType().getText(),
+								name: assertDefinedValue(param.getName()),
+								type: type.getText(),
 								hasQuestionToken: param.isOptional() && !param.isRestParameter(),
 								isRestParameter: param.isRestParameter()
 							};
+
+							if (type.isClassOrInterface()) {
+								p.type = type.getSymbolOrThrow().getName();
+							}
 
 							return p;
 						});
@@ -150,14 +156,20 @@ export default async function generate(source: string, options?: any): Promise<v
 		const dtsFile = project.addExistingSourceFile(dts);
 		for (const generateClass of generateClasses) {
 			const { className, methods } = generateClass;
-			const dtsClass: ClassDeclaration = dtsFile.getClass(className);
+			const dtsClass: ClassDeclaration = assertDefinedValue(dtsFile.getClass(className));
 
 			for (const method of methods) {
-				while (dtsClass.getStaticMethod(method.name)) {
-					dtsClass.getStaticMethod(method.name).remove();
+				let staticMethod = dtsClass.getStaticMethod(method.name);
+				let instanceMethod = dtsClass.getInstanceMethod(method.name);
+
+				while (staticMethod) {
+					staticMethod.remove();
+					staticMethod = dtsClass.getStaticMethod(method.name);
 				}
-				while (dtsClass.getInstanceMethod(method.name)) {
-					dtsClass.getInstanceMethod(method.name).remove();
+
+				while (instanceMethod) {
+					instanceMethod.remove();
+					instanceMethod = dtsClass.getInstanceMethod(method.name);
 				}
 
 				dtsClass.addMethod(method);