+++ pages/index.ts
@@ -1,10 +1,10 @@
 import { Credentials } from '@qa/account-manager';
 import * as Debug from 'debug';
 import * as merge from 'deepmerge';
-import config from '../config';
 import account, { UserType } from '../utils/account';
 import Authorization from '../utils/authorization';
 import URL from '../utils/url';
+import { assertDefinedValue } from '../utils/assert-defined';
 
 const debug = Debug('@qa:yoda');
 const TIMEOUT: number = 15 * 1000;
@@ -34,7 +34,7 @@ class PageObject {
 	 * @override this
 	 * @type {string}
 	 */
-	public readonly name: string = null;
+	public readonly name: string | null;
 
 	/**
 	 * Авторизация
@@ -139,7 +139,7 @@ class PageObject {
 		const { features, scripts } = cache;
 
 		if (features.length) {
-			query.ftrs = features.concat(query.ftrs).join(' ').trim();
+			query.ftrs = features.concat(query.ftrs || []).join(' ').trim();
 		}
 
 		url = URL.format(url, query);
@@ -352,7 +352,7 @@ class PageObject {
 	/**
 	 * Перемещение курсора, работюащее и в фф, и в хроме
 	 */
-	moveToElement(locator?: string, xoffset?: number, yoffset?: number) {
+	moveToElement(locator: string, xoffset?: number, yoffset?: number) {
 		if (browser.desiredCapabilities.browserName === 'firefox') {
 			this.moveToObjectAction(locator, xoffset, yoffset);
 		} else {