+++ steps/api/internal-api.ts
@@ -1,8 +1,10 @@
 import * as InternalApi from '../../api/internal';
+import { UserEditOptions } from '../../api/internal/users/edit';
 import { PhoneStatus } from '../../api/internal/user/phones-state';
 import userProfileSet from '../../api/internal/user/profile-set';
 import authorization from '../../store/authorization';
 import phonesStore from '../../store/phones';
+import { assertDefinedValue } from '../../utils/assert-defined';
 
 export type PhoneStatusStep = PhoneStatus | 'in_remove_queue' | 'twofa';
 export type ExtraEmailStatusStep = 'ok' | 'too_young' | 'in_remove_queue';
@@ -19,14 +21,15 @@ export default class InternalApiSteps {
 	@step('Добавить телефон для пользователя "{email}" с типом "{type}"')
 	addPhone(storeIndex: number, email: string, type: PhoneStatusStep) {
 		const phone: string = phonesStore.getNumber(storeIndex);
-
-		const phoneId = InternalApi.usersPhonesAdd({
+		const body = InternalApi.usersPhonesAdd({
 			email,
 			phones: [{
 				phone,
 				mobile: true
 			}]
-		}).body[0];
+		}).body;
+
+		const phoneId = assertDefinedValue(body)[0];
 
 		let initialType: PhoneStatus;
 		switch (type) {
@@ -81,7 +84,68 @@ export default class InternalApiSteps {
 			value
 		});
 
-		return response.body.body;
+		return assertDefinedValue(response.body).body;
+	}
+
+	@step('Задать пользователю дату рождения {__result__}')
+	setUserBirthday(login: string, domain: string, date: Required<UserEditOptions>['birthday']): string {
+		const userToEdit = {
+			login,
+			domain,
+			birthday: {
+				day: date.day,
+				month: date.month,
+				year: date.year
+			}
+		};
+
+		InternalApi.usersEdit({
+			users: [userToEdit]
+		});
+
+		return `${date.day}.${date.month}.${date.year}`;
+	}
+
+	@step('Задать пользователю таймзону c id {timezoneId},{autodetect ? "" : " не"} определять автоматически')
+	setUserTimezone(login: string, domain: string, timezoneId: number, autodetect: boolean = false) {
+		const userToEdit = {
+			login,
+			domain,
+			timezone_autodetect: autodetect,
+			timezone_id: timezoneId
+		};
+
+		InternalApi.usersEdit({
+			users: [userToEdit]
+		});
+	}
+
+	@step('Задать пользователю {sex === "male" ? "мужской" : "женский"} пол')
+	setUserGender(login: string, domain: string, sex: 'male' | 'female') {
+		const userToEdit = {
+			login,
+			domain,
+			sex
+		};
+
+		InternalApi.usersEdit({
+			users: [userToEdit]
+		});
+	}
+
+	@step('Задать данные имя, фамилию, псевдоним и дату рождения')
+	setUserNamesAndBirthday(login: string, domain: string, data: Partial<UserEditOptions>) {
+		const userToEdit = {
+			login,
+			domain,
+			birthday: data.birthday,
+			name: data.name,
+			nick: data.nick
+		};
+
+		InternalApi.usersEdit({
+			users: [userToEdit]
+		});
 	}
 }